version 1.2 Beta 1 :
= WordPress 3.0 Compatibility =
* Compat : WP 3.0 elimination of page.php, edit-pages.php, page-new.php broke many aspects of page filtering
* Compat : Support RS Roles, Restrictions for Custom Post Types created via WP 2.9 / 3.0 framework
* Compat : Support RS Roles for Custom Taxonomies created via WP 2.9 / 3.0 framework
* Compat : WP 3.0 Multisite menu items had invalid link

= New Features =
* Feature : Ajax interface for group membership selection
* Feature : Group membership requests
* Feature : Group membership recommendations (2-tier membership moderation)

= Major Bug Fixes =
* BugFix : File Filtering was not imposed based on Post/Page Restrictions or Default Category Roles (also required Private visibility)
* BugFix : RS Restrictions and Roles were not applied to Sticky Posts
* BugFix : Attachment filenames with spaces, parenthesis and other special chars caused corrupt or ineffective .htaccess (possibly resulting in Internal Server Error)
* BugFix : Last blog paging link sometimes hidden when Hidden Content Teaser enabled (also caused WP-PageNavi conflict)
* BugFix : With Revisionary (or possibly other plugins) enabled, posts are inappropriately forced into default category in logged user cannot post there.
* BugFix : Custom calls to wp_dropdown_pages (in template or other plugin code) were sometimes filtered inappropriately

= Minor Bug Fixes =
* BugFix : When previewing a post, non-editors don't see Page or Post listings in sidebar / topbar
* BugFix : Recent Comments widget included comments on unreadable posts, with WP 2.9
* BugFix : Custom WP_PLUGIN_DIR was not supported
* BugFix : In Bulk Object Roles Edit forms, links to edit roles of individual object were broken
* BugFix : RS addition to wp-admin footer forced horizontal scroll bar in IE7
* BugFix : Role Basis settings (User Roles and Group Roles enable / disable) were hidden and unalterable
* BugFix : If Page Reader is enabled as an "Additional Object Role", Private Page Reader also remains captioned as "Page Reader"
* BugFix : If Post Reader is enabled as an "Additional Object Role", Private Post Reader also remains captioned as "Post Reader"
* BugFix : Bad edit link on User Profile where user is a Group Manager for specific group(s)
* BugFix : When scanning Posts/Pages for unregistered attachments, File Attachment Utility did not distinguish broken links

= Plugin Compatibility =
* Compat : WP-PageNavi - conflict with paging links, see above
* Compat : Amember - PHP Warning (array_diff_key) after importing users
* Compat : QTranslate - unparsed page titles in Page Parent dropdown
* Compat : Simple Section Nav - children of excluded pages bubbled up to the page menu
* Compat : Reveal IDs plugin wiped out "Groups" column in Edit Users page
* Compat : Role Scoper potentially wiped out other plugin custom columns on Edit Users page

= Other Changes =
* Change : Apply Excerpt Teaser Prefix,Suffix whenever excerpt, pre-more, or first X chars replace content, if SCOPER_FORCE_EXCERPT_SUFFIX is defined.
* Perf : Don't load and initialize Role Scoper on asynchronous dashboard feed calls (WP dev blog, etc.)

git-svn-id: https://plugins.svn.wordpress.org/role-scoper/trunk@238218 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ admin/attachments_utility.php
@@ -43,7 +43,8 @@ function scoper_attach_linked_uploads( $echo = false ) {
 			$num_bites++;
 		
 		$upload_path = $uploads['baseurl'];
-
+		$upload_dir = $uploads['basedir'];
+		
 		if ( $echo ) {
 			printf(__( "<strong>checking %s posts / pages...</strong>", 'scoper' ), $num_posts);
 			echo '<br /><br />';
@@ -85,6 +86,17 @@ function scoper_attach_linked_uploads( $echo = false ) {
 							if ( $echo ) {
 								//printf( _ x( '<span class="rs-brown">skipping unfilterable file in %1$s "%2$s":</span> %3$s', 'post_type, post_title, file_url', 'scoper' ), __(ucwords($row->post_type)), $row->post_title, $file_url);
 								printf( __( '<span class="rs-brown">skipping unfilterable file in %1$s "%2$s":</span> %3$s', 'scoper' ), __(ucwords($row->post_type)), $row->post_title, $file_url);
+								echo '<br /><br />';
+							}
+						
+							continue;
+						} 
+						
+						// make sure the linked file actually exists
+						if ( ! file_exists( str_replace( $upload_path, $upload_dir, $file_url ) ) ) {
+							if ( $echo ) {
+								//printf( _ x( '<span class="rs-brown">skipping unfilterable file in %1$s "%2$s":</span> %3$s', 'post_type, post_title, file_url', 'scoper' ), __(ucwords($row->post_type)), $row->post_title, $file_url);
+								printf( __( '<span class="rs-red">skipping missing file in %1$s "%2$s":</span> %3$s', 'scoper' ), __(ucwords($row->post_type)), $row->post_title, $file_url);
 								echo '<br /><br />';
 							}
